security: comprehensive CNSA 2.0 Level 5 hardening + multi-dimensional security audit fixes
CNSA 2.0 SHA-256 to SHA-512 upgrades (18 files):
- proactive_sharing: PSS commitment hashes
- distributed_session: device fingerprint blind index (HMAC-SHA512)
- siem: dedup hash
- secret_ceremony: key material fingerprint
- build_manifest: binary integrity hash
- db_ha: backup manifest hash
- ceremony: sync HMAC (tag size 32->64 bytes)
- threshold OPAQUE: HMAC + verification key derivation (v2 domain sep)
- authorize: PKCE blind index + auth code hash
- clients: client secret salt
- tokens: refresh token hash
- gateway server: X-Wing key fingerprint (v2 domain separator)
- admin routes: CSRF token HMAC + token revocation hash
- he_search: searchable encryption HMAC
- enclave: sealing hash

Protocol security fixes:
- verifier: constant-time audience comparison (ct_eq)
- verifier: remove REQUIRE_TOKEN_AUDIENCE env var bypass (hardcode mandatory)
- ceremony: use secure_time instead of SystemTime::now for cleanup

Host compromise resistance:
- startup_checks: add 11 missing env vars to SENSITIVE_ENV_VARS sanitization
- ratchet: remove RATCHET_KEK from environment immediately after consumption
- threshold_kek: use per-boot HMAC fingerprint instead of bare SHA-512(KEK) in logs
- opaque store: volatile-write zeros over ServerSetup struct memory on Drop
- event_streaming: zeroize old webhook secret before replacement
- orchestrator messages: add Debug redaction + Drop zeroize for password

Code quality fixes:
- admin routes: add Debug impl for RegisterUserRequest (redact password)
- hsm: replace unreachable!() with Err return in seal/unseal_with_hardware
- fido registration: use i64 instead of i32 for WebAuthn sign_count (prevent u32 truncation)
- threshold_kek: replace .unwrap() with .expect() with reason

+++ fido/src/registration.rs
@@ -359,7 +359,7 @@ impl PersistentCredentialStore {
         .bind(&cred.credential_id)
         .bind(&pubkey_enc)
         .bind(cred.user_id)
-        .bind(cred.sign_count as i32)
+        .bind(cred.sign_count as i64)
         .bind(&cred.authenticator_type)
         .execute(&self.pool.pool)
         .await
@@ -402,7 +402,7 @@ impl PersistentCredentialStore {
     /// Update sign count in both DB and memory after successful authentication.
     pub async fn update_sign_count(&mut self, credential_id: &[u8], new_count: u32) -> Result<(), String> {
         sqlx::query("UPDATE fido_credentials SET sign_count = $1 WHERE credential_id = $2")
-            .bind(new_count as i32)
+            .bind(new_count as i64)
             .bind(credential_id)
             .execute(&self.pool.pool)
             .await
